consistent export of null statsd client
