+++ test/null.js
@@ -2,10 +2,10 @@
 
 var test = require('tape');
 
-var NullStatsd = require('../null.js');
+var createNullStatsd = require('../null.js');
 
 test('null.gauge()', function t(assert) {
-    var c = NullStatsd();
+    var c = createNullStatsd();
     c.gauge('some.key', 10);
 
     assert.deepEqual(c._buffer.peek(), {
@@ -21,7 +21,7 @@ test('null.gauge()', function t(assert) {
 });
 
 test('null.counter()', function t(assert) {
-    var c = NullStatsd();
+    var c = createNullStatsd();
     c.counter('some.key', 5);
 
     assert.deepEqual(c._buffer.peek(), {
@@ -37,7 +37,7 @@ test('null.counter()', function t(assert) {
 });
 
 test('null.increment()', function t(assert) {
-    var c = NullStatsd();
+    var c = createNullStatsd();
 
     c.increment('some.key');
 
@@ -65,7 +65,7 @@ test('null.increment()', function t(assert) {
 });
 
 test('null.decrement()', function t(assert) {
-    var c = NullStatsd();
+    var c = createNullStatsd();
 
     c.decrement('some.key');
 
@@ -104,7 +104,7 @@ test('null.decrement()', function t(assert) {
 });
 
 test('null.timing()', function t(assert) {
-    var c = NullStatsd();
+    var c = createNullStatsd();
 
     c.timing('some.key', 500);
